feat: Support responding with 401 status in middleware
Currently the middleware creates an error only with a 403 status.
This change supports providing "challenges" as a middleware option
in order to support 401 error responses.

A 401 error `status` may be created if the request is unauthorized
and the user role matches a key in the provided `challenges` option.
When the error is handled, the `challenges` value
should be used to determine the value of a `WWW-Authenticate`
response header.
I