Respond with a 401 status when access is denied for an anonymous user
When an anonymous client receives a 403 response, it can be confusing and difficult to distinguish between lack of permissions and lack of identity. This change makes the response error use a 401 status code and a different response text if the user role is anonymous. This is technically a breaking change and so includes a major version bump.